Fix numpy float

+++ nead/nead.py
@@ -62,9 +62,9 @@ def read(neadfile, MKS=None, multi_index=True, index_col=None):
 
             # Convert from string to number if it is a number
             if val.strip('-').strip('+').replace('.','').isdigit():
-                val = np.float(val)
-                if val == np.int(val):
-                    val = np.int(val)
+                val = float(val)
+                if val == int(val):
+                    val = int(val)
 
             if section == 'meta': meta[key] = val
             if section == 'fields': fields[key] = val
@@ -96,9 +96,9 @@ def read(neadfile, MKS=None, multi_index=True, index_col=None):
         if all([str(s).strip('-').strip('+').replace('.','').isdigit() or str(s) == "" for s in arr]):
             arr = np.array(arr).astype("<U32")
             arr[arr == ""] = 'nan'
-            arr = arr.astype(np.float)
-            if all(arr == arr.astype(np.int)):
-                arr = arr.astype(np.int)
+            arr = arr.astype(float)
+            if all(arr == arr.astype(int)):
+                arr = arr.astype(int)
                     
         for i,v in enumerate(ds.data_vars):
             # print(i,v)
